authorization basics

+++ routes/web.php
@@ -14,21 +14,22 @@ use App\Http\Controllers\IdeaController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
-//Whenever you get a request to this url pass it to index method controller
+    //Whenever you get a request to this url pass it to index method controller
 Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
 
 Route::post('/ideas', [IdeaController::class, 'store'])->name('ideas.store');
 
 Route::get('/ideas/{idea}', [IdeaController::class, 'show'])->name('ideas.show');
 
-Route::get('/ideas/{idea}/edit', [IdeaController::class, 'edit'])->name('ideas.edit');
+    // ->middleware('auth') ensures that user is logged in for that action
+Route::get('/ideas/{idea}/edit', [IdeaController::class, 'edit'])->name('ideas.edit')->middleware('auth');
 
-Route::put('/ideas/{idea}', [IdeaController::class, 'update'])->name('ideas.update');
+Route::put('/ideas/{idea}', [IdeaController::class, 'update'])->name('ideas.update')->middleware('auth');
 
-Route::delete('/ideas/{idea}', [IdeaController::class, 'destroy'])->name('ideas.destroy');
+Route::delete('/ideas/{idea}', [IdeaController::class, 'destroy'])->name('ideas.destroy')->middleware('auth');
 
 // - comments
-Route::post('/ideas/{idea}/comments', [CommentController::class, 'store'])->name('ideas.comments.store');
+Route::post('/ideas/{idea}/comments', [CommentController::class, 'store'])->name('ideas.comments.store')->middleware('auth');
 
 // - users
 Route::get('/register', [AuthController::class, 'register'])->name('register');
@@ -37,7 +38,7 @@ Route::post('/register', [AuthController::class, 'store']);
 Route::get('/login', [AuthController::class, 'login'])->name('login');
 Route::post('/login', [AuthController::class, 'authenticate']);
 
-Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
+Route::post('/logout', [AuthController::class, 'logout'])->name('logout')->middleware('auth');
 
 Route::get('/terms', function () {
     return view('terms');
